deposit and expenses working for total balance, styled all the pages, fixed any errors

+++ BudgetWebApp/src/main/java/com/example/demo/CustomUserDetails.java
@@ -62,7 +62,7 @@ public class CustomUserDetails implements UserDetails {
         return user.getId();
     }
 
-    public int getBalance(){
+    public double getBalance(){
         return user.getBalance();
     }
 }
